Select country through the HDX API
