+++ kuwala/pipelines/population-density/requirements.txt
@@ -1,2 +1,82 @@
+affine==2.3.0
+attrs==21.2.0
+basicauth==0.4.1
+beautifulsoup4==4.9.3
+blessed==1.17.6
+boto3==1.17.112
+botocore==1.20.112
+cchardet==2.1.7
+certifi==2021.5.30
+cffi==1.14.6
+chardet==4.0.0
+charset-normalizer==2.0.2
+ckanapi==4.6
+click==8.0.1
+click-plugins==1.1.1
+cligj==0.7.2
+colorlog==5.0.1
+cryptography==3.4.7
+decorator==5.0.9
+dnspython==2.1.0
+docopt==0.6.2
+email-validator==1.1.3
+et-xmlfile==1.1.0
+greenlet==1.1.0
+hdx-python-api==5.1.7
+hdx-python-country==2.8.8
+hdx-python-utilities==2.6.7
+html5lib==1.1
+idna==3.2
+ijson==3.1.4
+inflect==5.3.0
+itsdangerous==2.0.1
+jmespath==0.10.0
+jsonlines==2.0.0
+jsonpath-ng==1.5.3
+libhxl==4.24.1
+linear-tsv==1.1.0
+ndg-httpsclient==0.5.1
+num2words==0.5.10
+numpy==1.21.0
+openpyxl==3.0.7
+pandas==1.3.0
+ply==3.11
+prompt-toolkit==3.0.19
 py4j==0.10.9
+pyasn1==0.4.8
+pycparser==2.20
+pyOpenSSL==20.0.1
+pyparsing==2.4.7
+pyphonetics==0.5.3
 pyspark==3.1.2
+python-dateutil==2.8.1
+python-editor==1.0.4
+python-io-wrapper==0.2
+python-slugify==5.0.2
+pytz==2021.1
+PyYAML==5.4.1
+quantulum3==0.7.7
+questionary==1.10.0
+rasterio==1.2.6
+ratelimit==2.2.1
+readchar==2.0.1
+requests==2.26.0
+requests-cache==0.7.1
+requests-file==1.5.1
+ruamel.yaml==0.17.10
+ruamel.yaml.clib==0.2.6
+s3transfer==0.4.2
+six==1.16.0
+snuggs==1.4.7
+soupsieve==2.2.1
+SQLAlchemy==1.4.21
+tabulator==1.53.5
+text-unidecode==1.3
+unicodecsv==0.14.1
+Unidecode==1.2.0
+url-normalize==1.4.3
+urllib3==1.26.6
+wcwidth==0.2.5
+webencodings==0.5.1
+xlrd==2.0.1
+xlrd3==1.1.0
